requirements.txt
customtkinter
fastf1
matplotlib
numpy
pandas
timple
Pillow
seaborn
Flask
